+++ src/CurrentFileVersionCompatibilityTests/App.config
