+++ src/CurrentFileVersionCompatibilityTests/CurrentFileVersionCompatibilityTests.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{DCD83A54-54DA-4131-A852-2A129DB3ACB8}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>OldFileVersionWriteTests</RootNamespace>
-    <AssemblyName>OldFileVersionWriteTests</AssemblyName>
+    <RootNamespace>CurrentFileVersionCompatibilityTests</RootNamespace>
+    <AssemblyName>CurrentFileVersionCompatibilityTests</AssemblyName>
     <TargetFrameworkVersion>v4.7.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
