+++ src/CurrentFileVersionCompatibilityTests/Program.cs
@@ -4,13 +4,12 @@
 using Axe.Windows.Actions;
 using Axe.Windows.Actions.Contexts;
 using Axe.Windows.Core.Bases;
-using Axe.Windows.Core.Enums;
 using Axe.Windows.Core.Results;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace OldFileVersionWriteTests
+namespace CurrentFileVersionCompatibilityTests
 {
     class Program
     {
