+++ src/CurrentFileVersionCompatibilityTests/Properties/AssemblyInfo.cs
