+++ src/CurrentFileVersionCompatibilityTests/packages.config
